misc/azure-things/azure-data-things/container_test/pyproject.toml
[tool.poetry]
name = "container-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4"
azure-storage-blob = "12.20.0"
azure-identity = "1.17.1"

[tool.poetry.group.dev.dependencies]
pytest = "8.2.2"
ruff = "0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
